pogoapi/requirements.txt
flask
flask-cors
flask-socketio
python-dotenv
pymongo
pydantic
langchain
langchain-openai
langchain-core
langchain-anthropic
langchain-community
qdrant-client
pillow
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
requests
python-slugify
werkzeug>=2.0.0
gunicorn
eventlet
cryptography
marshmallow
marshmallow-dataclass
flask-jwt-extended
flask-marshmallow
flask-sqlalchemy
flask-migrate
alembic
langchain_google_genai
langchain_anthropic
python-docx
pyotp
celery==5.3.4
redis==5.0.1
flower==2.0.1
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
pyannote.audio
torch
firebase-admin
flask-mail